requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
build==1.4.0
cachetools==6.2.6
certifi==2026.2.25
charset-normalizer==3.4.5
chromadb==1.5.5
click==8.1.8
coloredlogs==15.0.1
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
exceptiongroup==1.3.1
filelock==3.19.1
flatbuffers==25.12.19
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
googleapis-common-protos==1.73.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.7.1
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
narwhals==2.18.0
numpy==2.0.2
oauthlib==3.3.1
onnxruntime==1.19.2
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.5
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.5
pyarrow==21.0.0
pybase64==1.4.3
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rpds-py==0.27.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.2
toml==0.10.2
tomli==2.4.0
tornado==6.5.5
tqdm==4.67.3
typer==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
zipp==3.23.0
